requirements.txt
tensorflow==2.14.0
keras==2.14.0
numpy==1.25.2
pandas==2.1.2
matplotlib==3.8.0
seaborn==0.13.0
scikit-learn==1.3.0
jupyter==1.0.0
opencv-python==4.8.0
Pillow==9.6.1